pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "The Brownfield Cartographer: codebase intelligence for rapid FDE onboarding."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "networkx>=3.2",
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "rich>=13.7",
  "sqlglot>=25.0",
  "tree-sitter>=0.22",
  "tree-sitter-languages>=1.10.2",
  "typer>=0.12",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
cartographer = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv]
dev-dependencies = [
  "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py310"
